Make tag order irrelevant
Previously, searching for tags was sensitive to the order of the tags.
i.e. searching for ":tag1:tag2:" would not return a heading that had
tags specified as ":tag2:tag1:".  It could be worked around by searching
for, e.g. ":tag1: :tag2:", but this behavior was non-obvious and
counter-intuitive.  Now strings containing tags are split into separate
tokens, making order irrelevant.