Match negations against whole outline path
This seems like the most sensible thing to do.  A possible downside is
that, when there are negations, it now requires getting the full outline
path for every node, whereas before that would only be done if
helm-org-rifle-show-path was on; so it may affect performance, but I'm
guessing not noticeably in most cases, and if there are no negations, it
won't matter.