Rewrite input parsing
Note that running delete-trailing-whitespace broke the tests, because it
modified the test data in the test.el file that had trailing whitespace
on some lines.  It took a while to track down why the tests were
failing, where the whitespace was coming from.  I'm not sure if I ran my
indent-whole-buffer function, or if it was my before-save-hook.